Export test setup helper functions
Moving them to the utils package was the original plan, although
it is not possible due to cyclic imports from engine and config.

Moving it to engine didn't make much sense to me, so I decided to
follow one of the recommendations from the official documentation
and create an internal package that itself contains packages
dedicated to implement one functionality. This could also replace
the utils package in the future.

An added benefit is that the testutil package will be ignored when
compiling our binaries due to it not being used anywhere except
in tests.

Link to docs: https://go.dev/doc/modules/layout
